refactored more to allow for use in other applications, but maintain the standalone option
